fix: cmake clean-up
