+++ external/CMakeLists.txt
@@ -3,5 +3,3 @@ add_subdirectory(eigen)
 add_subdirectory(cilantro)
 
 add_subdirectory(alglib)
-
-add_subdirectory(boost)
